% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dddSoilDischarge.init.R
\name{init}
\alias{init}
\title{init
The funtion initializes the soil dischareg
Unit is in m3/s}
\usage{
init(MAD, q1, D, Timeresinsec, modelArea, modelLayer, modelRiver, modelBog,
  layerUH, NoL, UHriver)
}
\arguments{
\item{MAD}{Mean Annual Discharge values}

\item{q1}{first value of the runoff timeserie}

\item{D}{volume of the unsaturated zone}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{modelArea}{list of parameters about the area
list(totarea=a,slopesriverarea=b,nobognoglacarea=c,bogarea=d)}

\item{modelLayer}{list of parameters about the Layers
list(maxL=a,speed=b,nbStepsDelay=c,z=d,distr="dexp",param=c(e), NoL=f)}

\item{modelRiver}{list of parameters about the river
list(maxL=a,speed=b,nbStepsDelay=c ,z=d,distr="dnorm",param=c(e,f))}

\item{modelBog}{list of parameters of the bog
list(maxL=a,speed=b,nbStepsDelay=c,z=d,distr="dexp",param=c(e))}

\item{layerUH}{Unit Hydrograph of the saturation layers}

\item{NoL}{number of level zone}

\item{UHriver}{Unit Hydrograp of the river}
}
\description{
init
The funtion initializes the soil dischareg
Unit is in m3/s
}
\examples{
\dontrun{
init()
}
}
\keyword{soilDischarge}

