% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.soilDischarge.R
\name{init.soilDischarge}
\alias{init.soilDischarge}
\title{Initializing the soil discharge
The funtion initializes the soil discharge
Unit is in m3/s}
\usage{
init.soilDischarge(method = NULL, path = NULL, qsimutx = NULL,
  qsimX = NULL, MAD = NULL, q1 = NULL, D = NULL, Timeresinsec = NULL,
  modelArea = NULL, modelLayer = NULL, modelRiver = NULL,
  modelBog = NULL, layerUH = NULL, UHriver = NULL)
}
\arguments{
\item{method}{method for the initialization, "load", "source", "manual", "processed"}

\item{path}{directory where to get the files, in used when method is "load" or "source"}

\item{qsimutx}{soil discharge at time t}

\item{qsimX}{accumulated soil discharge}

\item{MAD}{Mean Annual Discharge values}

\item{q1}{first value of the runoff timeserie}

\item{D}{volume of the unsaturated zone}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{modelArea}{list of parameters about the area
list(totarea,slopesriverarea,nobognoglacarea,bogarea)}

\item{modelLayer}{list of parameters about the Layers
list(maxL,speed,nbStepsDelay,z,distr,param,NoL)}

\item{modelRiver}{list of parameters about the river
list(maxL,speed,nbStepsDelay,z,distr,param)}

\item{modelBog}{list of parameters of the bog
list(maxL,speed,nbStepsDelay,z,dist,param)}

\item{layerUH}{Unit Hydrograph of the saturation layers}

\item{UHriver}{Unit Hydrograp of the river}
}
\description{
Initializing the soil discharge
The funtion initializes the soil discharge
Unit is in m3/s
}
\examples{
\dontrun{
init.soilDischarge()
}
}
\keyword{soilDischarge}

