% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdis.bogs.R
\name{sdis.bogs}
\alias{sdis.bogs}
\title{soil discharge on bogs}
\usage{
sdis.bogs(Timeresinsec, UHriver, modelBog, waterContent, area)
}
\arguments{
\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{UHriver}{Unit Hydrograp of the river}

\item{modelBog}{list of parameters of the bog
list(maxL=a,speed=b,nbStepsDelay=c,z=d,distr="dexp",param=c(e))}

\item{waterContent}{excess of water in millimeters}

\item{area}{in squared meters}
}
\description{
The function \code{sdis.bogs()} computes soil discharge on bogs routed through the river
}
\examples{
\dontrun{
sdis.bogs()
}
}
\keyword{soilDischarge}

