% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdis.D.R
\name{sdis.D}
\alias{sdis.D}
\title{Volume of unsaturated zone}
\usage{
sdis.D(q, Timeresinsec, area)
}
\arguments{
\item{q}{discharge in m3/s}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{area}{in squared meters}
}
\value{
A value in meters
}
\description{
The function \code{sdis.D()} computes the volume of unsaturared zone
}
\examples{
\dontrun{
dddDischarge.D()
}
}
\keyword{dddDischarge}

