% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdis.q.R
\name{sdis.q}
\alias{sdis.q}
\title{discharge}
\usage{
sdis.q(waterContent, area, Timeresinsec, UHxConvRiver)
}
\arguments{
\item{waterContent}{excess of water in millimeters}

\item{area}{in squared meters}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{UHxConvRiver}{convolution of the Unit Hydrograph of x to the Unit Hydrograph of the river}
}
\value{
a discharge value in m3/s
}
\description{
The function \code{sdis.q()} computes the discharge from a terrain feature routed through the river.
}
\examples{
\dontrun{
sdis.q()
}
}
\keyword{dddDischarge}

