% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdis.slopes.R
\name{sdis.slopes}
\alias{sdis.slopes}
\title{soil discharge on slopes}
\usage{
sdis.slopes(Timeresinsec, layerUH, ddist, UHriver, waterContent, area)
}
\arguments{
\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{layerUH}{Unit Hydrograph of the saturation layers}

\item{ddist}{states of each saturation level}

\item{UHriver}{Unit Hydrograp of the river}

\item{waterContent}{excess of water in millimeters}

\item{area}{in squared meters}
}
\description{
The function \code{sdis.slopes()} computes soil discharge on slopes routed through the river
}
\examples{
\dontrun{
sdis.slopes()
}
}
\keyword{soilDischarge}

